Fixed Spawning bug

+++ theSwarm/Larva.java
@@ -27,8 +27,8 @@ public class Larva {
 
                 if (nearByEnemies.length > 0)
                 {
-                    int[] AllEnemyBots = FightMicro.AllEnemyBots(rc);
-                    int[] AllEnemyNoiseTowers = FightMicro.AllEnemyNoiseTowers(rc);
+                    long[] AllEnemyBots = FightMicro.AllEnemyBots(rc);
+                    long[] AllEnemyNoiseTowers = FightMicro.AllEnemyNoiseTowers(rc);
 
                     FightMicro.FindAndRecordAllEnemies(rc, nearByEnemies, AllEnemyBots, AllEnemyNoiseTowers);
                 }
